+++ src/main/java/com/alexecollins/docker/mojo/ValidateMojo.java
@@ -1,10 +1,6 @@
 package com.alexecollins.docker.mojo;
 
-import java.util.logging.Level;
-import java.util.logging.Logger;
-
 import com.alexecollins.docker.orchestration.DockerOrchestrator;
-
 import org.apache.maven.plugins.annotations.LifecyclePhase;
 import org.apache.maven.plugins.annotations.Mojo;
 
@@ -18,18 +14,8 @@ import org.apache.maven.plugins.annotations.Mojo;
 @Mojo(name = "validate", defaultPhase = LifecyclePhase.VALIDATE)
 public class ValidateMojo extends AbstractDockerMojo {
 
-    private Logger logger = Logger.getLogger(ValidateMojo.class.getName());
-
     @Override
     protected void doExecute(DockerOrchestrator orchestrator) throws Exception {
-        logger.log(Level.INFO, "Starting validate DockerFile");
-        try {
-            orchestrator.validate();
-        } catch (Exception ex)
-        {
-            logger.log(Level.SEVERE, "invalid DockerFile", ex);
-            throw ex;
-        }
-        logger.log(Level.INFO, "DockerFile validated");
+        orchestrator.validate();
     }
 }